refactor: move main SD struct out of config into discovery/config
